Refactor transactions tiles
