--- src/orders/components/OrderManualTransationRefundPage/components/OrderTransactionTile/index.ts
@@ -1,1 +1,0 @@
-export * from "./OrderTransactionTile";
